+++ test/OpenTelemetry.Instrumentation.AWS.Tests/Implementation/AWSServiceHelperTests.cs
@@ -12,7 +12,7 @@ namespace OpenTelemetry.Instrumentation.AWS.Tests.Implementation;
 public class AWSServiceHelperTests
 {
     [Fact]
-    public void GetDynamoDbCloudRegion_RegionEndpointIsSet_CloudRegionIsCorrect()
+    public void ExtractCloudRegion_RegionEndpointIsSet_CloudRegionIsCorrect()
     {
         var region = "eu-central-1";
         var clientConfig = new AmazonDynamoDBConfig
@@ -21,7 +21,7 @@ public class AWSServiceHelperTests
         };
         var requestContext = new MockRequestContext(clientConfig);
 
-        var cloudRegion = AWSServiceHelper.GetDynamoDbCloudRegion(requestContext);
+        var cloudRegion = AWSServiceHelper.ExtractCloudRegion(requestContext);
 
         Assert.Equal(region, cloudRegion);
     }
@@ -29,7 +29,7 @@ public class AWSServiceHelperTests
     [Theory]
     [InlineData("https://dynamodb.eu-central-1.amazonaws.com")]
     [InlineData("https://a1b2c3.execute-api.eu-central-1.amazonaws.com/default/mydynamodb")]
-    public void GetDynamoDbCloudRegion_ServiceUrlIsSet_CloudRegionIsCorrect(string serviceUrl)
+    public void ExtractCloudRegion_ServiceUrlIsSet_CloudRegionIsCorrect(string serviceUrl)
     {
         var clientConfig = new AmazonDynamoDBConfig
         {
@@ -37,13 +37,13 @@ public class AWSServiceHelperTests
         };
         var requestContext = new MockRequestContext(clientConfig);
 
-        var cloudRegion = AWSServiceHelper.GetDynamoDbCloudRegion(requestContext);
+        var cloudRegion = AWSServiceHelper.ExtractCloudRegion(requestContext);
 
         Assert.Equal("eu-central-1", cloudRegion);
     }
 
     [Fact]
-    public void GetDynamoDbCloudRegion_RegionIsMissingInServiceUrl_CloudRegionIsNull()
+    public void ExtractCloudRegion_RegionIsMissingInServiceUrl_CloudRegionIsNull()
     {
         var clientConfig = new AmazonDynamoDBConfig
         {
@@ -51,22 +51,25 @@ public class AWSServiceHelperTests
         };
         var requestContext = new MockRequestContext(clientConfig);
 
-        var cloudRegion = AWSServiceHelper.GetDynamoDbCloudRegion(requestContext);
+        var cloudRegion = AWSServiceHelper.ExtractCloudRegion(requestContext);
 
-        Assert.Null(cloudRegion);
+        Assert.Equal("us-east-1", cloudRegion); // AWS SDK default
     }
 
-    [Fact]
-    public void GetDynamoDbCloudRegion_RegionNotFollowingEndpointPattern_CloudRegionIsReturned()
+    [Theory]
+    [InlineData("https://foo.bar.mycompany.com")]
+    [InlineData("https://foo.bar.mycompany.com/test")]
+    [InlineData("https://foo.baz.bar.mycomany.com/aaaa/bbbb")]
+    public void ExtractCloudRegion_RegionNotFollowingEndpointPattern_CorrectPartOfUrlReturned(string serviceUrl)
     {
         var clientConfig = new AmazonDynamoDBConfig
         {
-            ServiceURL = "https://dynamodb.us-west-new.amazonaws.com",
+            ServiceURL = serviceUrl,
         };
         var requestContext = new MockRequestContext(clientConfig);
 
-        var cloudRegion = AWSServiceHelper.GetDynamoDbCloudRegion(requestContext);
+        var cloudRegion = AWSServiceHelper.ExtractCloudRegion(requestContext);
 
-        Assert.Equal("us-west-new", cloudRegion);
+        Assert.Equal("us-east-1", cloudRegion); // AWS SDK default
     }
 }